added clustering algorithm

+++ app/Http/Controllers/UserController.php
@@ -23,8 +23,7 @@ class UserController extends Controller
             $event_id = $_GET['event_id'];
             $event = Event::find($event_id);
             $user = User::find($token);
-            if(count($event->members) == $event->max_allowed)
-                return 0;
+
             foreach ($event->members as $member) {
                 if($member->id == $token) {
                     $event->members()->detach($user);
@@ -33,6 +32,8 @@ class UserController extends Controller
                     return 2;
                 }
             }
+            if(count($event->members) == $event->max_allowed)
+                return 0;
             $event->members()->save($user);
             if(file_exists('places.txt'))
                 unlink('places.txt');
